Finished the core triggers for Screaming Fist
I finished implementing the key triggers that are needed for proper game
flow of The Screaming Fist. The last step to be completed will be to
balance combat and it's proper waves. Then this episode is ready for
re-release while I work on fixing the arcade/office section of The
Chase.